Add type cast long->size_t

+++ src/lib/unistd.c
@@ -116,7 +116,7 @@ int lfp_execvpe(const char *file, char *const argv[], char *const envp[])
     if (strchr(file, '/'))
         return execve(file, argv, envp);
 
-    long name_max = pathconf(file, _PC_NAME_MAX);
+    size_t name_max = (size_t) pathconf(file, _PC_NAME_MAX);
     size_t filelen = lfp_strnlen(file, name_max);
     SYSCHECK(ENAMETOOLONG, filelen >= name_max);
